separate modules into subfolders
