+++ main.py
@@ -5,9 +5,9 @@ import pycnbi.utils.pycnbi_utils as pu
 from PyQt5 import QtWidgets
 
 from package.entity.edata import constants, variables
-from package.views import main_view_new
-from package.views.stream_selector_view import StreamSelectorView
-from package.views.record_replay_selector import RecordReplaySelector
+from package.views.main_GUI import main_view
+from package.views.record_replay_GUI.record_replay_selector import RecordReplaySelector
+from package.views.stream_selector_GUI.stream_selector_view import StreamSelectorView
 
 
 # def display_views():
@@ -33,6 +33,6 @@ if __name__ == "__main__":
     record_replay_view = RecordReplaySelector()
     stream_selector_view = StreamSelectorView()
     # ex = main_view.MainView(variables.Variables.get_amp_name(), variables.Variables.get_amp_serial())
-    ex = main_view_new.MainView(variables.Variables.get_amp_name(), variables.Variables.get_amp_serial())
+    ex = main_view.MainView(variables.Variables.get_amp_name(), variables.Variables.get_amp_serial())
 
     sys.exit(app.exec_())
