+++ package/views/main_GUI/control_switches/main_switch.py
