+++ package/views/main_GUI/control_switches/scope_switch.py
