+++ package/views/main_GUI/exp_protocol_design/sequence_manager.py
