+++ package/views/main_GUI/exp_protocol_design/ssvep_exp_protocol_seperate.py
