+++ package/views/main_GUI/online_monitor/bad_epoch_monitor.py
