+++ package/views/main_GUI/oscilloscope/channel_scale_manager.py
