+++ package/views/main_GUI/oscilloscope/channel_selector.py
