+++ package/views/main_GUI/oscilloscope/event_plot.py
