+++ package/views/stream_selector_GUI/stream_selector_view.py
